+++ src/Schema/Directives/MutationExecutorDirective.php
@@ -33,18 +33,13 @@ abstract class MutationExecutorDirective extends BaseDirective implements FieldR
                 $modelClass = $this->getModelClass();
                 $model = new $modelClass();
 
-                $executeMutation = function () use ($model, $resolveInfo): Model {
-                    /** @var \Illuminate\Database\Eloquent\Model $mutated */
-                    $mutated = $this->executeMutation(
-                        $model,
-                        $resolveInfo->argumentSet
-                    );
-
-                    return $mutated->refresh();
-                };
-
                 return $this->transactionalMutations->execute(
-                    $executeMutation,
+                    function () use ($model, $resolveInfo): Model {
+                        $mutated = $this->executeMutation($model, $resolveInfo->argumentSet);
+                        assert($mutated instanceof Model);
+
+                        return $mutated->refresh();
+                    },
                     $model->getConnectionName()
                 );
             }
@@ -63,12 +58,12 @@ abstract class MutationExecutorDirective extends BaseDirective implements FieldR
             // Use the name of the argument if no explicit relation name is given
             ?? $this->nodeName();
 
-        /** @var \Illuminate\Database\Eloquent\Relations\Relation $relation */
         $relation = $parent->{$relationName}();
+        assert($relation instanceof Relation);
 
-        /** @var \Illuminate\Database\Eloquent\Model $related */
         // @phpstan-ignore-next-line Relation&Builder mixin not recognized
         $related = $relation->make();
+        assert($related instanceof Model);
 
         return $this->executeMutation($related, $args, $relation);
     }