+++ src/Schema/Types/LaravelEnumType.php
@@ -6,6 +6,7 @@ use BenSampo\Enum\Enum;
 use GraphQL\Type\Definition\Directive;
 use GraphQL\Type\Definition\EnumType;
 use InvalidArgumentException;
+use ReflectionClassConstant;
 
 /**
  * A convenience wrapper for registering enums programmatically.
@@ -32,10 +33,12 @@ class LaravelEnumType extends EnumType
      */
     public function __construct(string $enumClass, ?string $name = null)
     {
+        if (! class_exists($enumClass)) {
+            throw self::classDoesNotExist($enumClass);
+        }
+
         if (! is_subclass_of($enumClass, Enum::class)) {
-            throw new InvalidArgumentException(
-                "Must pass an instance of \BenSampo\Enum\Enum, got {$enumClass}."
-            );
+            throw self::classMustExtendBenSampoEnumEnum($enumClass);
         }
 
         $this->enumClass = $enumClass;
@@ -60,10 +63,22 @@ class LaravelEnumType extends EnumType
         ]);
     }
 
+    public static function classDoesNotExist(string $enumClass): InvalidArgumentException
+    {
+        return new InvalidArgumentException("Class {$enumClass} does not exist.");
+    }
+
+    public static function classMustExtendBenSampoEnumEnum(string $enumClass): InvalidArgumentException
+    {
+        $baseClass = Enum::class;
+
+        return new InvalidArgumentException("Class {$enumClass} must extend {$baseClass}.");
+    }
+
     protected function deprecationReason(Enum $enum): ?string
     {
         $constant = $this->reflection->getReflectionConstant($enum->key);
-        assert($constant instanceof \ReflectionClassConstant, 'Enum keys are derived from the constant names');
+        assert($constant instanceof ReflectionClassConstant, 'Enum keys are derived from the constant names');
 
         $docComment = $constant->getDocComment();
         if (false === $docComment) {