+++ tests/Unit/Subscriptions/Storage/RedisStorageManagerTest.php
@@ -164,6 +164,8 @@ class RedisStorageManagerTest extends TestCase
                     serialize($subscriber2),
                     // Simulate an expired key, see https://github.com/nuwave/lighthouse/issues/2035
                     null,
+                    // Depending on the setup, redis can also return this invalid result https://github.com/nuwave/lighthouse/issues/2085
+                    '(nil)',
                 ]
             );